api/requirements.txt
fastapi
uvicorn
pyjwt
SQLAlchemy
psycopg2-binary